fix formatting

+++ proteobench/io/parsing/parse_denovo.py
@@ -357,6 +357,7 @@ def _load_custom(input_path: str) -> pd.DataFrame:
     """
     return pd.read_csv(input_path, low_memory=False)
 
+
 def _load_peaks(input_path: str) -> pd.DataFrame:
     """
     Load a de novo output file in PEAKS format.
@@ -373,12 +374,13 @@ def _load_peaks(input_path: str) -> pd.DataFrame:
     """
     df = pd.read_csv(input_path, low_memory=False)
     # Drop unmappable entries
-    df = df[df['Scan']!='-']
-    
-    df['local confidence (%)'] = df["local confidence (%)"].apply(lambda x: [float(i) for i in x.split()])
-    df['ALC (%)'] = df['ALC (%)'].astype(float)
-    df['Scan'] = df['Scan'].astype(int)
+    df = df[df["Scan"] != "-"]
+
+    df["local confidence (%)"] = df["local confidence (%)"].apply(lambda x: [float(i) for i in x.split()])
+    df["ALC (%)"] = df["ALC (%)"].astype(float)
+    df["Scan"] = df["Scan"].astype(int)
     return df
+
 
 _LOAD_FUNCTIONS = {
     "AdaNovo": _load_adanovo,
